Valgrind cleanup. Made experience log treat it's exp list records properly and made sure m_exp_log is initialized in the constructor always. Also changed to use QPtrList since QList is deprecated.
Made spells.cpp properly call clear() which deletes the spells properly instead
of calling clearSpells (which is the signal when spells are cleared).

Made spawn.cpp initialize all spawn fields in all constructors.

Made xmlprefernces delete the sections when it is destructed.

Made spellshell not clobber spells if they are referenced twice.